requirements.txt
apeye>=0.4.0
attrs>=20.3.0
click>=7.1.2
consolekit>=0.6.0
domdf-python-tools>=1.7.0
dulwich>=0.20.13
flake8>=3.8.4
flake8-json>=19.8.0
requests>=2.25.0
typing-extensions>=3.7.4.3
